server/requirement.txt
apturl==0.5.2
asn1crypto==0.24.0
astroid==1.6.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
click==8.0.4
cmake==3.22.3
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defer==1.0.6
devscripts===2.17.12ubuntu1.1
distro==1.7.0
distro-info===0.18ubuntu0.18.04.1
easyocr==1.4.2
Flask==2.0.3
gitdb2==2.0.3
GitPython==2.1.8
gpg==1.10.0
gunicorn==20.1.0
httplib2==0.9.2
idna==2.6
imageio==2.15.0
importlib-metadata==4.8.3
isort==4.3.4
itsdangerous==2.0.1
Jinja2==3.0.3
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
logilab-common==1.4.1
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
netifaces==0.10.4
networkx==2.5.1
numpy==1.19.5
oauth==1.0.1
olefile==0.45.1
opencv-python-headless==4.5.4.60
packaging==21.3
paho-mqtt==1.5.1
pandas==1.1.5
pexpect==4.2.1
Pillow==8.4.0
protobuf==3.0.0
psutil==5.4.2
PyBluez==0.23
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.26.1
pylint==1.8.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==3.0.7
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.7
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian==0.1.32
python-magic==0.4.16
python-xapp==1.0.0
pytz==2018.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-build==0.14.1
scikit-image==0.17.2
scipy==1.5.4
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.13.2
six==1.11.0
smmap2==2.0.3
stdeb==0.10.0
stem==1.8.0
system-service==0.3
systemd-python==234
tifffile==2020.9.3
torch==1.10.1
torchvision==0.11.2
typing_extensions==4.1.1
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.4
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
Werkzeug==2.0.3
wrapt==1.9.0
xkit==0.0.0
xlrd==2.0.1
zipp==3.6.0
zope.interface==4.3.2
